Update ids.py formatting style
Done in an attempt to improve the readability of the IDs. This commit
does not include any changes to thresholds or bitmaps, nor does it add
any documentation.

Changes still need to be made for the updated electron and photon ID
definitions, as discussed in Issue #73. Documentation for Tau ID bitmaps
should also be added.