Fix formatting issues in ids.py
When reformatting ids.py to be more legible previously, the parentheses
that were removed inbetween conditional statements with comparisons
cannot be removed to preserve the intended behavior.

In Python, the bitwise operators & and | are executed before comparison
operators like <, ==, etc., which is not the intended behavior. On the
other hand, the boolean operators "and" and "or" are executed after
comparison operators as intended, but these operators are not overloaded
for this context. (See "Operator Precedence" in the offical Python
documentation).

Bitwise NOT ~ does come before & and |, and therefore does not require
the additional parentheses in these functions.

Some addtional formatting tweaks were also made to ids.py.